Change imap literal counts to parse and store unsigned ints.
IMAP literals are of type number.  Change imap_get_literal_count() to
use mutt_atoui() instead of atoi().  Change the return type variables
used to store the count to type unsigned int.

It's doubtful this was a real issue, but as long as we're cleaning up
incorrect atoi() usage, we should fix this too.

+++ imap/command.c
@@ -412,7 +412,7 @@ static void cmd_parse_list(struct ImapData *idata, char *s)
   struct ImapList *list = NULL;
   struct ImapList lb;
   char delimbuf[5]; /* worst case: "\\"\0 */
-  long litlen;
+  unsigned int litlen;
 
   if (idata->cmddata && idata->cmdtype == IMAP_CT_LIST)
     list = (struct ImapList *) idata->cmddata;
@@ -638,7 +638,7 @@ static void cmd_parse_status(struct ImapData *idata, char *s)
   unsigned int count;
   struct ImapStatus *status = NULL;
   unsigned int olduv, oldun;
-  long litlen;
+  unsigned int litlen;
   short new = 0;
   short new_msg_count = 0;
